Refactor: many changes to the object model of Reko.Core
* Moved src\Decompiler\IDecompiledFileService.cs to src\Decompiler\Services
* Moved src\Core\Services\IDecompilerUIService.cs to src\Gui\Services
* Sprinkled the [MaybeNullWhen] attribute on `out` parameters
* Removed one of the ExpressionEmitter.Slice overloads
* Updated the PublicAPI.Unshipped.txt file (still > 2000 API's to complete)

+++ src/Arch/Pdp11/Pdp11Rewriter.Alu.cs
@@ -191,7 +191,7 @@ namespace Reko.Arch.Pdp11
             SetFlags(quotient, Registers.NZVC);
         }
 
-        private void RewriteIncDec(Func<Expression, int, Expression> fn)
+        private void RewriteIncDec(Func<Expression, long, Expression> fn)
         {
             var src = RewriteSrc(instr.Operands[0]);
             var dst = RewriteDst(instr.Operands[0], src, s => fn(s, 1));
